fastai/requirements.txt
attrs==19.3.0
backcall==0.2.0
bcolz==1.2.1
bleach==3.1.5
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
graphviz==0.14
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isoweek==1.3.3
jedi==0.17.1
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
my-custom-code==1.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.19.0
opencv-python==4.2.0.34
packaging==20.4
pandas==1.0.5
pandas-summary==0.0.7
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pip==20.1.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
scikit-learn==0.20.4
sklearn-pandas==1.8.0
scipy==1.5.0
seaborn==0.10.1
Send2Trash==1.5.0
setuptools==40.8.0
six==1.15.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.34.2
widgetsnbextension==3.5.1
zipp==3.1.0